refactor(#792): move runtime_dsp inline tests to a sibling file (sweep)
runtime_dsp.rs was 979 lines — ~875 of them an inline #[cfg(test)] mod tests
(the #496 numeric-audit output-stage tests), which also violated the repo's
'tests live in <module>_tests.rs' rule. Moved the module to runtime_dsp_tests.rs
via #[cfg(test)] #[path] mod (alongside the existing lufs_tests). Production
drops 979 -> 106 and leaves the debt list. Pure move — engine tests (663) pass
from the new location, zero warnings.
